+++ app/Http/Controllers/Api/BaseApiController.php
@@ -20,6 +20,7 @@ use App\Transformer\Api\{
     AccountTransformer,
     NotificationTransformer,
     MediaTransformer,
+    MediaDraftTransformer,
     StatusTransformer
 };
 use League\Fractal;
@@ -192,7 +193,7 @@ class BaseApiController extends Controller
         ]);
     }
 
-    public function showTempMedia(Request $request, int $profileId, $mediaId)
+    public function showTempMedia(Request $request, int $profileId, $mediaId, $timestamp)
     {
         abort_if(!$request->user(), 403);
         abort_if(!$request->hasValidSignature(), 404); 
@@ -257,10 +258,9 @@ class BaseApiController extends Controller
         $media->save();
 
         $url = URL::temporarySignedRoute(
-            'temp-media', now()->addHours(1), ['profileId' => $profile->id, 'mediaId' => $media->id]
+            'temp-media', now()->addHours(1), ['profileId' => $profile->id, 'mediaId' => $media->id, 'timestamp' => time()]
         );
 
-        $preview_url = $url;
         switch ($media->mime) {
             case 'image/jpeg':
             case 'image/png':
@@ -279,7 +279,7 @@ class BaseApiController extends Controller
 
         $resource = new Fractal\Resource\Item($media, new MediaTransformer());
         $res = $this->fractal->createData($resource)->toArray();
-        $res['preview_url'] = $preview_url;
+        $res['preview_url'] = $url;
         $res['url'] = $url;
         return response()->json($res);
     }
@@ -308,8 +308,9 @@ class BaseApiController extends Controller
 
     public function verifyCredentials(Request $request)
     {
-        abort_if(!$request->user(), 403);
-        $id = Auth::id();
+        $user = $request->user();
+        abort_if(!$user, 403);
+        $id = $user->id;
 
         $res = Cache::remember('user:account:id:'.$id, now()->addHours(6), function() use($id) {
             $profile = Profile::whereNull('status')->whereUserId($id)->firstOrFail();
@@ -319,5 +320,20 @@ class BaseApiController extends Controller
         });
 
         return response()->json($res);
+    }
+
+    public function drafts(Request $request)
+    {
+        $user = $request->user();
+        abort_if(!$request->user(), 403);
+
+        $medias = Media::whereUserId($user->id)
+            ->whereNull('status_id')
+            ->latest()
+            ->take(13)
+            ->get();
+        $resource = new Fractal\Resource\Collection($medias, new MediaDraftTransformer());
+        $res = $this->fractal->createData($resource)->toArray();
+        return response()->json($res, 200, [], JSON_PRETTY_PRINT|JSON_UNESCAPED_SLASHES);
     }
 }